Refactor parser: extract to common project
