+++ common/src/main/scala/hmda/parser/filing/lar/ApplicantFormatValidator.scala
@@ -1,10 +1,10 @@
 package hmda.parser.filing.lar
 
-import hmda.model.filing.lar.{Applicant, Ethnicity, Race, Sex}
 import hmda.model.filing.lar.enums._
+import hmda.model.filing.lar.{Applicant, Ethnicity, Race, Sex}
+import hmda.parser.LarParserValidationResult
 import hmda.parser.filing.lar.LarParserErrorModel._
 import cats.implicits._
-import hmda.parser.LarParserValidationResult
 
 sealed trait ApplicantFormatValidator extends LarParser {
