+++ common/src/main/scala/hmda/parser/filing/lar/LarCsvParser.scala
@@ -2,7 +2,7 @@ package hmda.parser.filing.lar
 
 import hmda.model.filing.lar.LoanApplicationRegister
 import hmda.parser.ParserErrorModel.ParserValidationError
-import hmda.parser.filing.lar.LarFormatValidator._
+import hmda.parser.filing.lar.LarFormatValidator.validateLar
 
 object LarCsvParser {
   def apply(s: String)
